app/requirements.txt
openai>=1.30.0
chromadb>=0.5.5
python-dotenv>=1.0.1
rich>=13.7.0
tiktoken>=0.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0